Finish refactoring the Chain Sync Supplier
